requirements.txt
asgiref==3.2.7
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.5
django-cors-headers==3.2.1
django-import-export==2.1.0
django-simpleui==4.0.2
djangorestframework==3.11.0
et-xmlfile==1.0.1
idna==2.9
importlib-metadata==1.6.0
jdcal==1.4.1
jieba==0.42.1
lxml==4.5.0
MarkupPy==1.14
more-itertools==8.3.0
odfpy==1.4.1
openpyxl==3.0.3
packaging==20.4
pdfminer==20191125
pdfminer3k==1.3.1
Pillow==7.1.2
Pinyin2Hanzi==0.1.1
pluggy==0.13.1
ply==3.11
psycopg2==2.8.5
py==1.8.1
pyasn1==0.4.8
pycryptodome==3.9.7
pyparsing==2.4.7
pypinyin==0.37.0
pytest==5.4.2
python-docx==0.8.10
pytz==2019.3
pywin32==227
PyYAML==5.3.1
requests==2.23.0
rsa==4.0
six==1.15.0
sqlparse==0.3.1
tablib==1.1.0
urllib3==1.25.9
wcwidth==0.1.9
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
